tests/wpt/tests/trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests-wildcard.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  //No name given test
  test(t => {
    let policy = window.trustedTypes.createPolicy('SomeName', { createHTML: s => s } );
    assert_equals(policy.name, 'SomeName');
  }, "Wildcard given - policy creation works");
</script>
